etl/airflow/requirements.txt
yfinance==0.2.48
psycopg2-binary==2.9.9
sqlAlchemy
pandas